# fastapi/测试脚本/v2测试脚本/test/导出接口文档脚本/fastapi_complete_api_spec.json
{
...
  "openapi": "3.1.0",
  "info": {
    "title": "用户系统API",
    "description": "模块化的用户管理系统",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v2/md_editor/sessions": {
      "post": {
        "tags": [
          "v2-md_editor"
        ],
        "summary": "Create Session",
        "description": "创建编辑器会话",
        "operationId": "create_session_api_v2_md_editor_sessions_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditorSessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "v2-md_editor"
        ],
        "summary": "Get Sessions",
        "description": "获取用户的编辑器会话列表",
        "operationId": "get_sessions_api_v2_md_editor_sessions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "跳过的记录数",
              "default": 0,
              "title": "Skip"
            },
            "description": "跳过的记录数"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "返回的记录数",
              "default": 20,
              "title": "Limit"
            },
            "description": "返回的记录数"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EditorSessionResponse"
                  },
                  "title": "Response Get Sessions Api V2 Md Editor Sessions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/md_editor/sessions/{session_id}": {
      "get": {
        "tags": [
          "v2-md_editor"
        ],
        "summary": "Get Session",
        "description": "获取编辑器会话详情",
        "operationId": "get_session_api_v2_md_editor_sessions__session_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditorSessionDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "v2-md_editor"
        ],
        "summary": "Update Session",
        "description": "更新编辑器会话",
        "operationId": "update_session_api_v2_md_editor_sessions__session_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSessionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditorSessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "v2-md_editor"
        ],
        "summary": "Delete Session",
        "description": "删除编辑器会话",
        "operationId": "delete_session_api_v2_md_editor_sessions__session_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/md_editor/sessions/{session_id}/optimize": {
      "post": {
        "tags": [
          "v2-md_editor"
        ],
        "summary": "Optimize Content",
        "description": "AI优化内容",
        "operationId": "optimize_content_api_v2_md_editor_sessions__session_id__optimize_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptimizeContentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimizeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/md_editor/sessions/{session_id}/apply-optimization/{optimization_id}": {
      "post": {
        "tags": [
          "v2-md_editor"
        ],
        "summary": "Apply Optimization",
        "description": "应用AI优化结果",
        "operationId": "apply_optimization_api_v2_md_editor_sessions__session_id__apply_optimization__optimization_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          },
          {
            "name": "optimization_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Optimization Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/md_editor/sessions/{session_id}/save-document": {
      "post": {
        "tags": [
          "v2-md_editor"
        ],
        "summary": "Save As Document",
        "description": "将编辑器会话保存为正式文档",
        "operationId": "save_as_document_api_v2_md_editor_sessions__session_id__save_document_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveDocumentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/md_editor/sessions/{session_id}/optimizations": {
      "get": {
        "tags": [
          "v2-md_editor"
        ],
        "summary": "Get Session Optimizations",
        "description": "获取会话的AI优化历史",
        "operationId": "get_session_optimizations_api_v2_md_editor_sessions__session_id__optimizations_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AIOptimizationResponse"
                  },
                  "title": "Response Get Session Optimizations Api V2 Md Editor Sessions  Session Id  Optimizations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/md_editor/stats": {
      "get": {
        "tags": [
          "v2-md_editor"
        ],
        "summary": "Get Editor Stats",
        "description": "获取编辑器统计信息",
        "operationId": "get_editor_stats_api_v2_md_editor_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditorStatsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v2/md_editor/config": {
      "get": {
        "tags": [
          "v2-md_editor"
        ],
        "summary": "Get Editor Config",
        "description": "获取编辑器配置",
        "operationId": "get_editor_config_api_v2_md_editor_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AIOptimizationResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "original_content": {
            "type": "string",
            "title": "Original Content"
          },
          "optimized_content": {
            "type": "string",
            "title": "Optimized Content"
          },
          "optimization_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Optimization Prompt"
          },
          "ai_provider": {
            "type": "string",
            "title": "Ai Provider"
          },
          "status": {
            "$ref": "#/components/schemas/OptimizationStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "original_content",
          "optimized_content",
          "optimization_prompt",
          "ai_provider",
          "status",
          "created_at"
        ],
        "title": "AIOptimizationResponse"
      },
...
      "CreateSessionRequest": {
        "properties": {
          "document_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "session_type": {
            "$ref": "#/components/schemas/SessionType",
            "default": "new_document"
          }
        },
        "type": "object",
        "title": "CreateSessionRequest"
      },
...
      "EditorSessionDetailResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "document_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "is_draft": {
            "type": "boolean",
            "title": "Is Draft"
          },
          "session_type": {
            "$ref": "#/components/schemas/SessionType"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "ai_optimizations": {
            "items": {
              "$ref": "#/components/schemas/AIOptimizationResponse"
            },
            "type": "array",
            "title": "Ai Optimizations",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "document_id",
          "title",
          "content",
          "is_draft",
          "session_type",
          "created_at",
          "updated_at"
        ],
        "title": "EditorSessionDetailResponse"
      },
      "EditorSessionResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "document_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "is_draft": {
            "type": "boolean",
            "title": "Is Draft"
          },
          "session_type": {
            "$ref": "#/components/schemas/SessionType"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "document_id",
          "title",
          "content",
          "is_draft",
          "session_type",
          "created_at",
          "updated_at"
        ],
        "title": "EditorSessionResponse"
      },
      "EditorStatsResponse": {
        "properties": {
          "total_sessions": {
            "type": "integer",
            "title": "Total Sessions"
          },
          "draft_sessions": {
            "type": "integer",
            "title": "Draft Sessions"
          },
          "total_optimizations": {
            "type": "integer",
            "title": "Total Optimizations"
          },
          "recent_sessions": {
            "items": {
              "$ref": "#/components/schemas/EditorSessionResponse"
            },
            "type": "array",
            "title": "Recent Sessions"
          }
        },
        "type": "object",
        "required": [
          "total_sessions",
          "draft_sessions",
          "total_optimizations",
          "recent_sessions"
        ],
        "title": "EditorStatsResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OptimizationStatus": {
        "type": "string",
        "enum": [
          "pending",
          "completed",
          "failed"
        ],
        "title": "OptimizationStatus"
      },
      "OptimizeContentRequest": {
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1,
            "title": "Content",
            "description": "要优化的MD内容"
          },
          "optimization_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Optimization Type",
            "description": "优化类型：general, grammar, structure, expand",
            "default": "general"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "OptimizeContentRequest"
      },
      "OptimizeResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "optimization_id": {
            "type": "integer",
            "title": "Optimization Id"
          },
          "original_content": {
            "type": "string",
            "title": "Original Content"
          },
          "optimized_content": {
            "type": "string",
            "title": "Optimized Content"
          },
          "can_apply": {
            "type": "boolean",
            "title": "Can Apply",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "success",
          "message",
          "optimization_id",
          "original_content",
          "optimized_content"
        ],
        "title": "OptimizeResponse"
      },
...
      "SaveDocumentRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "folder_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder Id"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "SaveDocumentRequest"
      },
      "SessionType": {
        "type": "string",
        "enum": [
          "new_document",
          "edit_document"
        ],
        "title": "SessionType"
      },
...
      "UpdateSessionRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "is_draft": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Draft"
          }
        },
        "type": "object",
        "title": "UpdateSessionRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}